Fixed error message
Fixed an error message that reports if the renamed Papyrus compiler
cannot be found in the same folder as the wrapper. The message was
previously saying that the renamed compiler could not be found in the
renamed compiler when it should have been saying that the renamed
compiler could not be found in the folder that should contain said
renamed compiler.